sqlliveness: move GC loop up to slsession, rename
This cleanup commit reworks which layer handles the GC loop.
It also removes the references to the name instance in the
sqlliveness layer.

Release note: None
