+++ pkg/sql/sqlliveness/slsession/BUILD.bazel
@@ -3,7 +3,9 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "slsession",
     srcs = [
-        "slinstance.go",
+        "factory.go",
+        "session.go",
+        "settings.go",
         "test_helpers.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/sql/sqlliveness/slsession",
@@ -12,6 +14,7 @@ go_library(
         "//pkg/settings",
         "//pkg/settings/cluster",
         "//pkg/sql/sqlliveness",
+        "//pkg/sql/sqlliveness/slstorage",
         "//pkg/util/hlc",
         "//pkg/util/log",
         "//pkg/util/retry",
@@ -19,12 +22,13 @@ go_library(
         "//pkg/util/syncutil",
         "//pkg/util/timeutil",
         "//pkg/util/uuid",
+        "@com_github_pkg_errors//:errors",
     ],
 )
 
 go_test(
     name = "slsession_test",
-    srcs = ["slinstance_test.go"],
+    srcs = ["factory_test.go"],
     deps = [
         ":slsession",
         "//pkg/clusterversion",
@@ -36,6 +40,8 @@ go_test(
         "//pkg/util/leaktest",
         "//pkg/util/log",
         "//pkg/util/stop",
+        "//pkg/util/timeutil",
+        "@com_github_cockroachdb_errors//:errors",
         "@com_github_stretchr_testify//require",
     ],
 )
