+++ pkg/sql/sqlliveness/slstorage/slstorage.go
@@ -12,8 +12,6 @@ package slstorage
 
 import (
 	"context"
-	"math/rand"
-	"time"
 
 	"github.com/cockroachdb/cockroach/pkg/keys"
 	"github.com/cockroachdb/cockroach/pkg/kv"
@@ -31,35 +29,9 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/util/stop"
 	"github.com/cockroachdb/cockroach/pkg/util/syncutil"
 	"github.com/cockroachdb/cockroach/pkg/util/syncutil/singleflight"
-	"github.com/cockroachdb/cockroach/pkg/util/timeutil"
 	"github.com/cockroachdb/errors"
 	"github.com/cockroachdb/logtags"
 	"github.com/cockroachdb/redact"
-)
-
-// GCInterval specifies duration between attempts to delete extant
-// sessions that have expired.
-var GCInterval = settings.RegisterDurationSetting(
-	"server.sqlliveness.gc_interval",
-	"duration between attempts to delete extant sessions that have expired",
-	20*time.Second,
-	settings.NonNegativeDuration,
-)
-
-// GCJitter specifies the jitter fraction on the interval between attempts to
-// delete extant sessions that have expired.
-//
-// [(1-GCJitter) * GCInterval, (1+GCJitter) * GCInterval]
-var GCJitter = settings.RegisterFloatSetting(
-	"server.sqlliveness.gc_jitter",
-	"jitter fraction on the duration between attempts to delete extant sessions that have expired",
-	.15,
-	func(f float64) error {
-		if f < 0 || f > 1 {
-			return errors.Errorf("%f is not in [0, 1]", f)
-		}
-		return nil
-	},
 )
 
 // CacheSize is the size of the entries to store in the cache.
@@ -73,22 +45,33 @@ var CacheSize = settings.RegisterIntSetting(
 	"number of session entries to store in the LRU",
 	1024)
 
+// Writer provides interactions with the storage of session records.
+type Writer interface {
+	// Insert stores the input Session.
+	Insert(ctx context.Context, id sqlliveness.SessionID, expiration hlc.Timestamp) error
+	// Update looks for a Session with the same SessionID as the input Session in
+	// the storage and if found replaces it with the input returning true.
+	// Otherwise it returns false to indicate that the session does not exist.
+	Update(ctx context.Context, id sqlliveness.SessionID, expiration hlc.Timestamp) (bool, error)
+	// DeleteExpiredSessions will delete all discovered sessions with an expiration less than now.
+	// Any errors encountered will be recorded into logging and metrics.
+	DeleteExpiredSessions(ctx context.Context, now hlc.Timestamp)
+}
+
 // Storage implements sqlliveness.Storage.
 type Storage struct {
-	settings   *cluster.Settings
-	stopper    *stop.Stopper
-	clock      *hlc.Clock
-	db         *kv.DB
-	codec      keys.SQLCodec
-	metrics    Metrics
-	gcInterval func() time.Duration
-	g          singleflight.Group
-	newTimer   func() timeutil.TimerI
-	tableID    descpb.ID
+	settings *cluster.Settings
+	stopper  *stop.Stopper
+	clock    *hlc.Clock
+	db       *kv.DB
+	codec    keys.SQLCodec
+	metrics  Metrics
+	g        singleflight.Group
+	tableID  descpb.ID
 
 	mu struct {
 		syncutil.Mutex
-		started bool
+
 		// liveSessions caches the current view of expirations of live sessions.
 		liveSessions *cache.UnorderedCache
 		// deadSessions caches the IDs of sessions which have not been found. This
@@ -109,7 +92,6 @@ func NewTestingStorage(
 	codec keys.SQLCodec,
 	settings *cluster.Settings,
 	sqllivenessTableID descpb.ID,
-	newTimer func() timeutil.TimerI,
 ) *Storage {
 	s := &Storage{
 		settings: settings,
@@ -118,14 +100,7 @@ func NewTestingStorage(
 		db:       db,
 		codec:    codec,
 		tableID:  sqllivenessTableID,
-		newTimer: newTimer,
-		gcInterval: func() time.Duration {
-			baseInterval := GCInterval.Get(&settings.SV)
-			jitter := GCJitter.Get(&settings.SV)
-			frac := 1 + (2*rand.Float64()-1)*jitter
-			return time.Duration(frac * float64(baseInterval.Nanoseconds()))
-		},
-		metrics: makeMetrics(),
+		metrics:  makeMetrics(),
 	}
 	cacheConfig := cache.Config{
 		Policy: cache.CacheLRU,
@@ -146,24 +121,12 @@ func NewStorage(
 	codec keys.SQLCodec,
 	settings *cluster.Settings,
 ) *Storage {
-	return NewTestingStorage(stopper, clock, db, codec, settings, keys.SqllivenessID,
-		timeutil.DefaultTimeSource{}.NewTimer)
+	return NewTestingStorage(stopper, clock, db, codec, settings, keys.SqllivenessID)
 }
 
 // Metrics returns the associated metrics struct.
 func (s *Storage) Metrics() *Metrics {
 	return &s.metrics
-}
-
-// Start runs the delete sessions loop.
-func (s *Storage) Start(ctx context.Context) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if s.mu.started {
-		return
-	}
-	_ = s.stopper.RunAsyncTask(ctx, "slstorage", s.deleteSessionsLoop)
-	s.mu.started = true
 }
 
 // IsAlive determines whether a given session is alive. If this method returns
@@ -185,10 +148,6 @@ func (s *Storage) isAlive(
 	ctx context.Context, sid sqlliveness.SessionID, syncOrAsync readType,
 ) (alive bool, _ error) {
 	s.mu.Lock()
-	if !s.mu.started {
-		s.mu.Unlock()
-		return false, sqlliveness.NotStartedError
-	}
 	if _, ok := s.mu.deadSessions.Get(sid); ok {
 		s.mu.Unlock()
 		s.metrics.IsAliveCacheHits.Inc(1)
@@ -309,32 +268,16 @@ func (s *Storage) deleteOrFetchSession(
 	return alive, expiration, nil
 }
 
-// deleteSessionsLoop is launched in start and periodically deletes sessions.
-func (s *Storage) deleteSessionsLoop(ctx context.Context) {
-	ctx, cancel := s.stopper.WithCancelOnQuiesce(ctx)
-	defer cancel()
-	t := s.newTimer()
-	t.Reset(s.gcInterval())
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-t.Ch():
-			t.MarkRead()
-			s.deleteExpiredSessions(ctx)
-			t.Reset(s.gcInterval())
-		}
-	}
-}
+// DeleteExpiredSessions is part of the Writer interface.
+func (s *Storage) DeleteExpiredSessions(ctx context.Context, now hlc.Timestamp) {
 
-// TODO(ajwerner): find a way to utilize this table scan to update the
-// expirations stored in the in-memory cache or remove it altogether. As it
-// stand, this scan will run more frequently than sessions expire but it won't
-// propagate that fact to IsAlive. It seems like the lazy session deletion
-// which has been added should be sufficient to delete expired sessions which
-// matter. This would closer align with the behavior in node-liveness.
-func (s *Storage) deleteExpiredSessions(ctx context.Context) {
-	now := s.clock.Now()
+	// TODO(ajwerner): find a way to utilize this table scan to update the
+	// expirations stored in the in-memory cache or remove it altogether. As it
+	// stand, this scan will run more frequently than sessions expire but it won't
+	// propagate that fact to IsAlive. It seems like the lazy session deletion
+	// which has been added should be sufficient to delete expired sessions which
+	// matter. This would closer align with the behavior in node-liveness.
+
 	var deleted int64
 	ctx = multitenant.WithTenantCostControlExemption(ctx)
 	if err := s.db.Txn(ctx, func(ctx context.Context, txn *kv.Txn) error {
